SQLEXEC-54: Add Rule Access Control for Tenant, Ruleset, and Schema APIs

+++ backend/app/api/routers/ruleset_manager.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 
 from api.core.services.ruleset.ruleset_manager_service import RulesetManagerService
 from model.requests.ruleset_manager.add_ruleset_request import AddRulesetRequest
@@ -9,17 +9,37 @@ from utils.jwt_utils import authenticate_admin_session
 router = APIRouter()
 
 @router.post("/{tenant_id}/ruleset")
-async def add_ruleset(tenant_id: str, ruleset_request: AddRulesetRequest, _: AdminSessionData = Depends(authenticate_admin_session)):
+async def add_ruleset(tenant_id: str, ruleset_request: AddRulesetRequest, admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Ruleset Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await RulesetManagerService.add_ruleset(tenant_id=tenant_id,ruleset_request=ruleset_request)
 
 @router.get("/{tenant_id}/ruleset/{ruleset_name}")
-async def get_ruleset(tenant_id: str, ruleset_name: str, _: AdminSessionData = Depends(authenticate_admin_session)):
+async def get_ruleset(tenant_id: str, ruleset_name: str, admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Ruleset Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await RulesetManagerService.get_ruleset(tenant_id=tenant_id, ruleset_name=ruleset_name)
 
 @router.put("/{tenant_id}/ruleset/{ruleset_name}")
-async def update_ruleset(tenant_id: str, ruleset_name: str, update_ruleset_request: AddRulesetRequest, _: AdminSessionData = Depends(authenticate_admin_session)):
+async def update_ruleset(tenant_id: str, ruleset_name: str, update_ruleset_request: AddRulesetRequest, admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Ruleset Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await RulesetManagerService.update_ruleset(tenant_id=tenant_id, ruleset_name=ruleset_name, update_ruleset_request=update_ruleset_request)
 
 @router.delete("/{tenant_id}/ruleset/{ruleset_name}")
-async def update_ruleset(tenant_id: str, ruleset_name: str, _: AdminSessionData = Depends(authenticate_admin_session)):
+async def update_ruleset(tenant_id: str, ruleset_name: str, admin_session_data: AdminSessionData = Depends(authenticate_admin_session)):
+    if admin_session_data.role not in ["Ruleset Admin", "Tenant Admin"]:
+        raise HTTPException(
+            status_code=403,
+            detail="You do not have permission to perform this action"
+        )
     return await RulesetManagerService.delete_ruleset(tenant_id=tenant_id,ruleset_name=ruleset_name)
